Rewriting of exfi.io - Renamed functions - Compartimentalized functions - Compartimentalized tests
