+++ bin/build_splice_graph
@@ -1,6 +1,15 @@
 #!/usr/bin/env python3
 
 import argparse
+
+from os.path import \
+    isfile, \
+    abspath
+
+from Bio import SeqIO
+
+from exfi import __version__
+
 from exfi.find_exons import \
     _get_fasta, \
     _find_exons_pipeline
@@ -12,14 +21,14 @@ from exfi.correct_splice_graph import \
     correct_splice_graph
 
 from exfi.io import \
-    index_fasta, \
-    write_gfa1
+    index_fasta
 
-from Bio import SeqIO
-from os.path import \
-    isfile, \
-    abspath
-from exfi import __version__
+from exfi.io.splice_graph_to_gfa1 import \
+    splice_graph_to_gfa1
+
+
+
+
 
 parser = argparse.ArgumentParser(
     usage='build_splicegraph -i transcriptome.fa -b bloom_filter.bf -k 30 1-o exome.gfa',
@@ -132,7 +141,7 @@ if __name__ == "__main__":
     splice_graph = correct_splice_graph(splice_graph, args)
 
     # Write to GFA1
-    write_gfa1(
+    splice_graph_to_gfa1(
         splice_graph=splice_graph,
         transcriptome_dict=transcriptome_dict,
         filename=args["output_gfa"]
